Place e OpenMap aggiornati

+++ VerificaWEB2/src/verificaweb2/VerificaWEB2.java
@@ -14,10 +14,12 @@ public class VerificaWEB2 {
 
     public static void main(String[] args) throws IOException, MalformedURLException, SAXException, ParserConfigurationException {
         Risposta r;
-        Place posti = new Place();
+        Place posto = new Place();
         OpenMap mappa = new OpenMap();
         Scanner myObj = new Scanner(System.in);  // Create a Scanner object
         Utente ut = new Utente();
+
+        boolean gandalf = true;
 
         int c = 0; // scelta
         System.out.println("\n0. Registrazione \n1. Login");
@@ -58,64 +60,67 @@ public class VerificaWEB2 {
                     ut.setToken(r.getDato());
                 } else {
                     System.out.println(r.getDato()); // <- Errore 
+                    gandalf = false;
                 }
 
                 break;
         }
 
-        do {
+        if (gandalf) {
+            do {
 
-            System.out.println("\n 1. Inserisci tappa \n 2. Visualizza lista tappe \n 3. Rimuovi tappa\n 4. Inverti tappe\n 5. Calcola distanza itinerario\n 6. Cancella itinerario\n 0. Esci");
-            c = myObj.nextInt();
+                System.out.println("\n 1. Inserisci tappa \n 2. Visualizza lista tappe \n 3. Rimuovi tappa\n 4. Inverti tappe\n 5. Calcola distanza itinerario\n 6. Cancella itinerario\n 0. Esci");
+                c = myObj.nextInt();
 
-            switch (c) {
-                case 1: // inserisci tappa
-                    //String tappa = setTappa();
-                    //mappa.run(tappa);
-                    r = WebService.setString(ut.getToken());
-                    //System.out.println(r.getDato());
+                switch (c) {
+                    case 1: // inserisci tappa
+                        String tappa = setTappa();
+                        posto = mappa.place(tappa);
 
-                    break;
-                case 2: // visualizza tappe
+                        r = WebService.setString(ut.getToken(), posto);
 
-                    WebService.getString(ut.getToken());
+                        break;
+                    case 2: // visualizza tappe
 
-                    break;
-                case 3: // rimuovi tappa
-                    
-                    System.out.println("Numero della tappa da togliere?");              
-                    
-                    int num = 0;
-                    
-                    num = myObj.nextInt();
-                    r = WebService.deleteString(ut.getToken(), num);
-                    
-                    if(r.getStato()){
-                        System.out.println("Tappa eliminata");
-                    } else {
-                        System.out.println("Errore"); 
-                    }
-                    
-                    break;
-                case 4: // inverti tappa
-                    // swap tra le due 
-                    
-                    int x = 0; 
-                    int y = 0; 
-                    
-                    break;
-                case 5: // calcola distanza itinerario 
-                    break;
-                case 6: // cancella itinerario 
-                    
-                    WebService.deleteAllStrings(ut.getToken());
-                                    
-                    break;
-                case 0: // esci 
-                    break;
-            }
-        } while (c != 0);
+                        WebService.getString(ut.getToken());
 
+                        break;
+                    case 3: // rimuovi tappa
+
+                        System.out.println("Numero della tappa da togliere?");
+
+                        int num = 0;
+
+                        num = myObj.nextInt();
+                        r = WebService.deleteString(ut.getToken(), num);
+
+                        if (r.getStato()) {
+                            System.out.println("Tappa eliminata");
+                        } else {
+                            System.out.println("Errore");
+                        }
+
+                        break;
+                    case 4: // inverti tappa
+                        // swap tra le due 
+
+                        int x = 0;
+                        int y = 0;
+
+                        break;
+                    case 5: // calcola distanza itinerario 
+
+                        break;
+                    case 6: // cancella itinerario 
+
+                        WebService.deleteAllStrings(ut.getToken());
+
+                        break;
+                    case 0: // esci 
+                        break;
+                }
+            } while (c != 0);
+        }
     }
 
     private static String setTappa() {
